backend/requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
google-auth-oauthlib==1.1.0
httpx==0.25.0
alembic==1.12.1
pydantic_settings
argon2-cffi==23.1.0
passlib[argon2]==1.7.4